backend reordering, minor front end re-orddering
