+++ Iwacu/backend/routes/posts.js
@@ -1,112 +1,23 @@
 const express = require("express");
-const multer = require("multer");
 
 const Post = require("../models/post");
 const checkAuth = require("../middleware/check-auth");
+const extractFile = require("../middleware/file")
+
+const postController = require("../controllers/posts");
 
 
 const router = express.Router();
 
-const MIME_TIME_MAP = {
-  'image/png': 'png',
-  'image/jpeg': 'jpg',
-  'image/jpg': 'jpg'
-}
+router.post("", checkAuth, extractFile, postController.createPost);
 
-const storage = multer.diskStorage({
-  destination: (req, file, cb)=> {
-    const isValid = MIME_TIME_MAP[file.mimetype];
-    let error = new Error ("Invalid mime type");
-    if(isValid){
-      error= null;
-    }
-    cb(error, "backend/images")
-  },
-  filename : (req, file, cb) => {
-    const name = file.originalname.toLowerCase().split(' ').join('-');
-    const ext = MIME_TIME_MAP[file.mimetype];
-    cb(null, name + '-' + Date.now() + '.' + ext);
-  }
-});
+router.put("/:id",  checkAuth, extractFile, postController.updatePost);
 
-router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
-  const url = req.protocol + '://' + req.get("host");
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename,
-    creator: req.userData.userId
-  });
-  post.save().then(createdPost => {
-    res.status(201).json({
-      message: "Post added successfully",
-      post: {
-        id : createdPost._id,
-        title: createdPost.title,
-        content: createdPost.content,
-        imagePath: createdPost.imagePath,
+router.get("", postController.getPosts);
 
-      }
-    });
-  });
-});
+router.get("/:id", postController.getPost);
 
-router.put("/:id",  checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
-  let imagePath = req.body.imagePath;
-  if(req.file){
-    const url = req.protocol + '://' + req.get("host");
-    imagePath = url + "/images/" + req.file.filename;
-  }
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath,
-    creator: req.userData.userId
-  });
-  Post.updateOne(
-    { _id: req.params.id, creator: req.userData.userId}, post
-    ).then(result => {
-      if(result.modifiedCount > 0){
-        res.status(200).json({ message: "Update successful!" });
-      } else {
-        res.status(401).json({ message: "Not Authorized!" });
-      }
-
-  });
-});
-
-router.get("", (req, res, next) => {
-  Post.find().then(documents => {
-    res.status(200).json({
-      message: "Posts fetched successfully!",
-      posts: documents
-    });
-  });
-});
-
-router.get("/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if (post) {
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({ message: "Post not found!" });
-    }
-  });
-});
-
-router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
-    result => {
-      console.log(result);
-      if (result.deletedCount > 0) {
-        res.status(200).json({ message: "Deletion successful!" });
-      } else {
-        res.status(401).json({ message: "Not authorized!" });
-      }
-    }
-  );
-});
+router.delete("/:id", checkAuth, postController.deletePost);
 
 
 module.exports = router;
