+++ Iwacu/src/app/angular-material.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from "@angular/core";
+import{ MatInputModule } from '@angular/material/input';
+import{ MatCardModule } from '@angular/material/card';
+import{ MatButtonModule} from '@angular/material/button'
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import {MatDialogModule} from '@angular/material/dialog';
+
+@NgModule({
+  exports: [
+    MatInputModule,
+    MatCardModule,
+    MatButtonModule,
+    MatToolbarModule,
+    MatExpansionModule,
+    MatProgressSpinnerModule,
+    MatPaginatorModule,
+    MatDialogModule,
+    MatGridListModule
+
+  ]
+})
+export class AngularMaterialModule {}
