+++ Iwacu/src/app/posts/post-list/post-list.component.ts
@@ -20,7 +20,7 @@ export class PostListComponent implements OnInit, OnDestroy {
   posts: Post[] = [];
   isLoading = false;
   totalPosts = 0;
-  postsPerPage = 2;
+  postsPerPage = 7;
   currentPage = 1;
   pageSizeOptions = [1, 2, 5, 10];
   userId: string;
